+++ CHANGELOG.md
@@ -35,7 +35,8 @@
   * upgraded to stubbles/core 4.x
   * added `net\stubbles\input\ValueReader::asSecureString()`
   * added `net\stubbles\input\filter\SecureStringFilter`
-  * added `net\stubbles\input\filter\range\StringLength` can also work with instances of `stubbles\lang\SecureString`
+  * `net\stubbles\input\filter\range\StringLength` can also work with instances of `stubbles\lang\SecureString`
+  * request broker now supports turning param values into instances of `stubbles\lang\SecureString` with `@Request[SecureString]`
 
 
 2.5.1 (2014-05-07)